requirements.txt
absl-py==1.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
celluloid==0.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
deepdish==0.3.7
defusedxml==0.7.1
eagerpy==0.30.0
easing-functions==1.0.4
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.2
filelock==3.7.1
fire==0.4.0
Flask==2.1.3
flatbuffers==1.12
fonttools==4.34.4
foolbox==3.3.3
gast==0.4.0
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.0
h5py==3.7.0
idna==3.3
imageio==2.20.0
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.1
lime==0.2.0.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==2.0.4
mtcnn==0.1.1
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.5
nets==0.0.3.1
networkx==2.8.5
nibabel==4.0.1
nilearn==0.9.1
nltk==3.2.4
nltools==0.4.5
notebook==6.4.12
numexpr==2.8.3
numpy==1.23.1
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
preprocessing==0.1.13
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-feat==0.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pynv==0.2
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
pyzmq==23.2.0
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.9.0
seaborn==0.11.2
Send2Trash==1.8.0
sewar==0.4.5
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
sphinx-rtd-theme==0.2.4
stack-data==0.3.0
tables==3.7.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.7.28
tinycss2==1.1.1
tldextract==3.3.1
torch==1.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.1
widgetsnbextension==4.0.3
wrapt==1.14.1
zipp==3.8.1
zope.interface==5.4.0
